rename package from expression to spdxexp
Thoughts on the renaming:
- expression is long for a go package name
- exp is not specific (e.g. exp.Satisfies)
- spdxexp hits the balance between short length and enough info to quickly distinguish it from other packages (e.g. spdxexp.Satisfies)
